Split dependencies-hierarchy into separate files
