+++ reviewing/dependencies-hierarchy/src/types.ts
@@ -1,0 +1,17 @@
+export type SearchFunction = (pkg: { name: string, version: string }) => boolean
+
+export interface PackageNode {
+  alias: string
+  circular?: true
+  dependencies?: PackageNode[]
+  dev?: boolean
+  isPeer: boolean
+  isSkipped: boolean
+  isMissing: boolean
+  name: string
+  optional?: true
+  path: string
+  resolved?: string
+  searched?: true
+  version: string
+}
